+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed TAS rendering popping up command prompt probably
 - Fixed music overlaying on each other every restart
+- Fixed UniTAS not loading at all for games
 
 ### Compatibility